Remove docker-compose-test.yml
