--- docker-compose-test.yml
@@ -1,169 +1,0 @@
-version: '3'
-
-services:
-  api:
-    build:
-      context: .
-      args: 
-        PYTHON_VERSION: ${PYTHON_VERSION}
-    container_name: lamapi_api
-    env_file:
-      - ./.env
-    command: bash -c "gunicorn -w $THREADS -b 0.0.0.0:5000 server:app --timeout 300 --reload --log-level debug"
-    ports:
-      - "${LAMAPI_PORT}:5000"
-    restart: always
-    volumes:
-      - ./api/:/app
-      - ./scripts/:/scripts
-  
-  mongo:
-    image: mongo:${MONGO_VERSION}
-    container_name: lamapi_mongo
-    env_file:
-      - ./.env
-    ports:
-      - "${MONGO_PORT}:27017"
-    restart: always   
-    volumes:
-      - ./mongo-data:/data/db
-      - ${VOLUME_MY_DATA_PATH}:/data/my-data  
-
-  es01:
-    image: docker.elastic.co/elasticsearch/elasticsearch:${STACK_VERSION}
-    volumes:
-      - esdata01:/usr/share/elasticsearch/data
-    ports:
-      - ${ELASTIC_PORT}:9200
-    environment:
-      - node.name=es01
-      - cluster.name=${CLUSTER_NAME}
-      - cluster.initial_master_nodes=es01,es02,es03
-      - discovery.seed_hosts=es02,es03
-      - bootstrap.memory_lock=true
-      - xpack.security.enabled=false
-      - xpack.license.self_generated.type=basic
-    mem_limit: ${MEM_LIMIT}
-    ulimits:
-      memlock:
-        soft: -1
-        hard: -1
-    healthcheck:
-      test:
-        [
-          "CMD-SHELL",
-          "curl -s http://localhost:9200 | grep -q 'You Know, for Search'",
-        ]
-      interval: 10s
-      timeout: 10s
-      retries: 120
-
-  es02:
-    depends_on:
-      - es01
-    image: docker.elastic.co/elasticsearch/elasticsearch:${STACK_VERSION}
-    volumes:
-      - esdata02:/usr/share/elasticsearch/data
-    environment:
-      - node.name=es02
-      - cluster.name=${CLUSTER_NAME}
-      - cluster.initial_master_nodes=es01,es02,es03
-      - discovery.seed_hosts=es01,es03
-      - bootstrap.memory_lock=true
-      - xpack.security.enabled=false
-      - xpack.license.self_generated.type=basic
-    mem_limit: ${MEM_LIMIT}
-    ulimits:
-      memlock:
-        soft: -1
-        hard: -1
-    healthcheck:
-      test:
-        [
-          "CMD-SHELL",
-          "curl -s http://localhost:9200 | grep -q 'You Know, for Search'",
-        ]
-      interval: 10s
-      timeout: 10s
-      retries: 120
-
-  es03:
-    depends_on:
-      - es02
-    image: docker.elastic.co/elasticsearch/elasticsearch:${STACK_VERSION}
-    volumes:
-      - esdata03:/usr/share/elasticsearch/data
-    environment:
-      - node.name=es03
-      - cluster.name=${CLUSTER_NAME}
-      - cluster.initial_master_nodes=es01,es02,es03
-      - discovery.seed_hosts=es01,es02
-      - bootstrap.memory_lock=true
-      - xpack.security.enabled=false
-      - xpack.license.self_generated.type=basic
-    mem_limit: ${MEM_LIMIT}
-    ulimits:
-      memlock:
-        soft: -1
-        hard: -1
-    healthcheck:
-      test:
-        [
-          "CMD-SHELL",
-          "curl -s http://localhost:9200 | grep -q 'You Know, for Search'",
-        ]
-      interval: 10s
-      timeout: 10s
-      retries: 120
-
-  kibana:
-    depends_on:
-      es01:
-        condition: service_healthy
-      es02:
-        condition: service_healthy
-      es03:
-        condition: service_healthy
-    image: docker.elastic.co/kibana/kibana:${STACK_VERSION}
-    ports:
-      - ${KIBANA_PORT}:5601
-    environment:
-      - SERVERNAME=kibana
-      - ELASTICSEARCH_HOSTS=http://es01:9200
-    mem_limit: ${MEM_LIMIT}
-    healthcheck:
-      test:
-        [
-          "CMD-SHELL",
-          "curl -s -I http://localhost:5601 | grep -q 'HTTP/1.1 302 Found'",
-        ]
-      interval: 10s
-      timeout: 10s
-      retries: 120
-
-  jupyter:
-    image: roby944/jupyter-extended
-    container_name: lamapi_jupyter
-    env_file:
-      - ./.env
-    environment:
-      - JUPYTER_ENABLE_LAB= "yes"
-    command: >
-      bash -c "jupyter lab --ip=0.0.0.0 --port=8888 --no-browser --allow-root --NotebookApp.token='' --NotebookApp.password=''"
-    ports:
-      - "${MY_JUPYTER_PORT}:8888"
-    restart: always   
-    volumes:
-      - ./work:/home/jovyan/work
-      - ./api:/home/jovyan/work/api
-      - ${VOLUME_MY_DATA_PATH}:/home/jovyan/work/my-data
-      - ./scripts:/home/jovyan/work/scripts
-      
-
-volumes:
-  esdata01:
-    driver: local
-  esdata02:
-    driver: local
-  esdata03:
-    driver: local
